4_imgModel/requirements.txt
absl-py==1.4.0
array-record==0.4.1
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
diffusers==0.21.4
dm-tree==0.1.8
etils==1.5.0
filelock==3.12.4
flatbuffers==23.5.26
fsspec==2023.9.2
ftfy==6.1.1
gast==0.5.4
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
grpcio==1.59.0
h5py==3.9.0
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
keras==2.14.0
keras-core==0.1.7
keras-cv==0.6.4
libclang==16.0.6
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ml-dtypes==0.2.0
namex==0.0.7
numpy==1.26.0
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
Pillow==10.0.1
promise==2.3
protobuf==3.20.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rsa==4.9
safetensors==0.4.0
six==1.16.0
stable-diffusion-tf==0.1
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-addons==0.21.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.14.0
termcolor==2.3.0
toml==0.10.2
tqdm==4.66.1
typeguard==2.13.3
typing_extensions==4.8.0
urllib3==2.0.6
wcwidth==0.2.8
Werkzeug==3.0.0
wrapt==1.14.1
zipp==3.17.0
# pytorch == 2.0.1